4.0.1

+++ provider/src/main/java/com/raygun/raygun4android/CrashReporting.java
@@ -29,7 +29,7 @@ public class CrashReporting {
     private static CrashReportingOnBeforeSend onBeforeSend;
     private static List tags;
     private static Map customData;
-    private static List<RaygunBreadcrumbMessage> breadcrumbs;
+    private static List<RaygunBreadcrumbMessage> breadcrumbs = new ArrayList<>();
     private static boolean shouldProcessBreadcrumbLocation = false;
 
     static RaygunUncaughtExceptionHandler getExceptionHandler() {
@@ -62,22 +62,44 @@ public class CrashReporting {
     }
 
     static void recordBreadcrumb(RaygunBreadcrumbMessage breadcrumb) {
-        breadcrumbs.add(processBreadcrumbLocation(breadcrumb, shouldProcessBreadcrumbLocation(),3));
+        breadcrumbs.add(processBreadcrumbLocation(breadcrumb, shouldProcessBreadcrumbLocation()));
     }
 
-    private static RaygunBreadcrumbMessage processBreadcrumbLocation(RaygunBreadcrumbMessage breadcrumb, boolean shouldProcessBreadcrumbLocation, int stackFrame) {
+    static void clearBreadcrumbs() {
+        breadcrumbs.clear();
+    }
+
+    private static RaygunBreadcrumbMessage processBreadcrumbLocation(RaygunBreadcrumbMessage breadcrumb, boolean shouldProcessBreadcrumbLocation) {
 
         if(shouldProcessBreadcrumbLocation && breadcrumb.getClassName() == null) {
-            StackTraceElement frame = Thread.currentThread().getStackTrace()[stackFrame];
 
-            return new RaygunBreadcrumbMessage.Builder(breadcrumb.getMessage())
-                .category(breadcrumb.getCategory())
-                .customData(breadcrumb.getCustomData())
-                .level(breadcrumb.getLevel())
-                .className(frame.getClassName())
-                .methodName(frame.getMethodName())
-                .lineNumber(frame.getLineNumber())
-                .build();
+            StackTraceElement[] trace = Thread.currentThread().getStackTrace();
+            StackTraceElement frame = null;
+
+            if (trace != null && trace.length > 0) {
+                for (int i = 0; i < trace.length-1; i++) {
+                    StackTraceElement thisFrame =trace[i];
+                    StackTraceElement nextFrame = trace[i+1];
+
+                    if (thisFrame.getClassName().contains("com.raygun.raygun4android.") && !nextFrame.getClassName().contains("com.raygun.raygun4android.")) {
+                        frame = nextFrame;
+                        break;
+                    }
+                }
+            }
+
+            if (frame != null) {
+                return new RaygunBreadcrumbMessage.Builder(breadcrumb.getMessage())
+                    .category(breadcrumb.getCategory())
+                    .customData(breadcrumb.getCustomData())
+                    .level(breadcrumb.getLevel())
+                    .className(frame.getClassName())
+                    .methodName(frame.getMethodName())
+                    .lineNumber(frame.getLineNumber())
+                    .build();
+            }
+
+
         }
 
         return breadcrumb;
@@ -163,31 +185,36 @@ public class CrashReporting {
     static void postCachedMessages() {
         if (RaygunNetworkUtils.hasInternetConnection(RaygunClient.getApplicationContext())) {
             File[] fileList = RaygunClient.getApplicationContext().getCacheDir().listFiles(new RaygunFileFilter());
-            for (File f : fileList) {
-                try {
-                    if (RaygunFileUtils.getExtension(f.getName()).equalsIgnoreCase(RaygunSettings.DEFAULT_FILE_EXTENSION)) {
-                        ObjectInputStream ois = null;
-                        try {
-                            ois = new ObjectInputStream(new FileInputStream(f));
-                            SerializedMessage serializedMessage = (SerializedMessage) ois.readObject();
-                            enqueueWorkForCrashReportingService(RaygunClient.getApiKey(), serializedMessage.message);
-                            if (!f.delete()) {
-                                RaygunLogger.w("Couldn't delete cached report (" + f.getName() + ")");
-                            }
-                        } finally {
-                            if (ois != null) {
-                                ois.close();
+            if (fileList != null) {
+                for (File f : fileList) {
+                    try {
+                        if (RaygunFileUtils.getExtension(f.getName()).equalsIgnoreCase(RaygunSettings.DEFAULT_FILE_EXTENSION)) {
+                            ObjectInputStream ois = null;
+                            try {
+                                ois = new ObjectInputStream(new FileInputStream(f));
+                                SerializedMessage serializedMessage = (SerializedMessage) ois.readObject();
+                                enqueueWorkForCrashReportingService(RaygunClient.getApiKey(), serializedMessage.message);
+                                if (!f.delete()) {
+                                    RaygunLogger.w("Couldn't delete cached report (" + f.getName() + ")");
+                                }
+                            } finally {
+                                if (ois != null) {
+                                    ois.close();
+                                }
                             }
                         }
+                    } catch (FileNotFoundException e) {
+                        RaygunLogger.e("Error loading cached message from filesystem - " + e.getMessage());
+                    } catch (IOException e) {
+                        RaygunLogger.e("Error reading cached message from filesystem - " + e.getMessage());
+                    } catch (ClassNotFoundException e) {
+                        RaygunLogger.e("Error in handling cached message from filesystem - " + e.getMessage());
                     }
-                } catch (FileNotFoundException e) {
-                    RaygunLogger.e("Error loading cached message from filesystem - " + e.getMessage());
-                } catch (IOException e) {
-                    RaygunLogger.e("Error reading cached message from filesystem - " + e.getMessage());
-                } catch (ClassNotFoundException e) {
-                    RaygunLogger.e("Error in cached message from filesystem - " + e.getMessage());
                 }
+            } else {
+                RaygunLogger.e("Error in handling cached message from filesystem - could not get a list of files from cache dir");
             }
+
         }
     }
